LoverCodes/__init__.py
from LoverCodes.core.bot import Lover
from LoverCodes.core.dir import dirr
from LoverCodes.core.git import git
from LoverCodes.core.userbot import Userbot
from LoverCodes.misc import dbb, heroku
from pyrogram import Client
from SafoneAPI import SafoneAPI
from .logging import LOGGER

dirr()
git()
dbb()
heroku()

app = Lover()
api = SafoneAPI()
userbot = Userbot()

from .platforms import *

Apple = AppleAPI()
Carbon = CarbonAPI()
SoundCloud = SoundAPI()
Spotify = SpotifyAPI()
Resso = RessoAPI()
Telegram = TeleAPI()
YouTube = YouTubeAPI()
