requirements.txt
aiofiles
aiohttp
apscheduler
beautifulsoup4
bs4
telegraph
wget
unidecode
dnspython
emojis
ffmpeg-python
geopy
gitpython
hachoir
heroku3
git+https://github.com/AsmSafone/SafoneAPI
img2pdf
langdetect
motor
numpy
pillow
psutil
pycountry
pycryptodome
pyfiglet
python-dotenv
#
python-whois
pytz
pyyaml==6.0.1
pykeyboard
python-telegram-bot
#httpx[http2]==0.27.2
httpx>=0.28.1
#httpx==0.27.0
qrcode
requests
speedtest-cli
stripe==10.7.0
telebot
git+https://github.com/TeamZyro/daxxhub.git
youtube-search-python
git+https://github.com/KurimuzonAkuma/pyrogram.git
git+https://github.com/yt-dlp/yt-dlp.git
py-tgcalls==2.2.11 
#git+https://github.com/MrZyro/ZyroTgCall.git
SpeechRecognition
spotipy
asyncio
bard
bing_image_downloader
lexica-api
Faker
instaloader
openai==0.28.0
pydub
pyyaml==6.0.1
MukeshAPI
youtube-search
gpytranslate
gtts
mysql-connector-python==9.0.0
search_engine_parser
opencv-python-headless
yt-search-py==1.0.0
